Added scroll toggle/minor details, fixed auto-embed
Minor details being adding an X to the flood control bubble, changed
animation to fade in/out, and auto fade out after 4 seconds if you
haven't clicked it yet.  I also worked on displaying the unread messages
with a border on the left hand side which fades out linearly from the
first to the last once you see it (as an indicator that these messages
are new).  Currently, I commented it out as I couldn't get the animation
to do exactly what I had in mind, but it's there for
possibility/improvement (would also need to be extended to
connect/disconnect messages).