Added ability to view images fully when clicked
Set max-height for images 150px with rounded edges.  When clicked, it
will expand to its full size in the center of the window if it can
comfortably be seen on screen.  Otherwise, you will have the additional
option of zooming in and out on the image from a comfortable size to
actual size.